digital-seal/backend/requirements.txt
fastapi
uvicorn
numpy
opencv-python
ImageHash
Pillow
python-multipart
requests
reedsolo
requests
# Structural similarity (SSIM) used for robust crop/derivative detection
scikit-image
# Ethereum wallet signature verification (MetaMask eth_sign / personal_sign)
eth-account
